+++ src/main.cpp
@@ -5,6 +5,7 @@
 #include "engine/SummaryDiff.h"
 #include "rules/DivByZeroRule.h"
 #include "rules/IntOverflowRule.h"
+#include "rules/SignConversionRule.h"
 #include "rules/BoundsRule.h"
 #include "rules/AssumptionRule.h"
 #include "rules/MemoryLeakRule_Ex.h"
@@ -64,6 +65,7 @@ int main(int argc, char* argv[]) {
     analyzer.addRule<codeskeptic::MemoryLeakRule_Ex>();
     analyzer.addRule<codeskeptic::DivByZeroRule>();
     analyzer.addRule<codeskeptic::IntOverflowRule>();
+    analyzer.addRule<codeskeptic::SignConversionRule>();
     analyzer.addRule<codeskeptic::BoundsRule>();
     analyzer.addRule<codeskeptic::AssumptionRule>();
     analyzer.addRule<codeskeptic::NullDerefRule>();