+++ src/server/McpServer.cpp
@@ -4,6 +4,7 @@
 #include "config/Config.h"
 #include "rules/DivByZeroRule.h"
 #include "rules/IntOverflowRule.h"
+#include "rules/SignConversionRule.h"
 #include "rules/BoundsRule.h"
 #include "rules/AssumptionRule.h"
 #include "rules/MemoryLeakRule_Ex.h"
@@ -169,6 +170,7 @@ json::Value runAnalyze(const json::Value& id, const json::Object* args) {
     analyzer.addRule<codeskeptic::MemoryLeakRule_Ex>();
     analyzer.addRule<codeskeptic::DivByZeroRule>();
     analyzer.addRule<codeskeptic::IntOverflowRule>();
+    analyzer.addRule<codeskeptic::SignConversionRule>();
     analyzer.addRule<codeskeptic::BoundsRule>();
     analyzer.addRule<codeskeptic::AssumptionRule>();
     analyzer.addRule<codeskeptic::NullDerefRule>();